+++ previous-versions/assess/codatassess.go
@@ -69,12 +69,12 @@ func (c *sdkConfiguration) GetServerDetails() (string, map[string]string) {
 //
 // [See our OpenAPI spec](https://github.com/codatio/oas)
 type CodatAssess struct {
-	// Match mutable accounting data with immutable banking data to increase confidence in financial data
-	DataIntegrity *dataIntegrity
-	// Downloadable reports
-	ExcelReports *excelReports
 	// Enriched reports and analyses of financial data
-	Reports *reports
+	Reports *Reports
+	// Match mutable accounting data with immutable banking data to increase confidence in financial data
+	DataIntegrity *DataIntegrity
+	// Downloadable reports
+	ExcelReports *ExcelReports
 
 	sdkConfiguration sdkConfiguration
 }
@@ -152,9 +152,9 @@ func New(opts ...SDKOption) *CodatAssess {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "1.0",
-			SDKVersion:        "0.26.0",
-			GenVersion:        "2.159.2",
-			UserAgent:         "speakeasy-sdk/go 0.26.0 2.159.2 1.0 github.com/codatio/client-sdk-go/previous-versions/assess",
+			SDKVersion:        "0.27.0",
+			GenVersion:        "2.195.2",
+			UserAgent:         "speakeasy-sdk/go 0.27.0 2.195.2 1.0 github.com/codatio/client-sdk-go/previous-versions/assess",
 		},
 	}
 	for _, opt := range opts {
@@ -173,11 +173,11 @@ func New(opts ...SDKOption) *CodatAssess {
 		}
 	}
 
+	sdk.Reports = newReports(sdk.sdkConfiguration)
+
 	sdk.DataIntegrity = newDataIntegrity(sdk.sdkConfiguration)
 
 	sdk.ExcelReports = newExcelReports(sdk.sdkConfiguration)
-
-	sdk.Reports = newReports(sdk.sdkConfiguration)
 
 	return sdk
 }