+++ previous-versions/assess/pkg/models/operations/getrecurringrevenuemetrics.go
@@ -31,8 +31,6 @@ func (o *GetRecurringRevenueMetricsRequest) GetConnectionID() string {
 type GetRecurringRevenueMetricsResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
-	ErrorMessage *shared.ErrorMessage
 	// OK
 	Report *shared.Report
 	// HTTP response status code for this operation
@@ -46,13 +44,6 @@ func (o *GetRecurringRevenueMetricsResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *GetRecurringRevenueMetricsResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *GetRecurringRevenueMetricsResponse) GetReport() *shared.Report {